refactor: Consolidate Linux sysfs helpers into linux/sysfs subpackage
Move IsSysPathRAID, GetSysPaths and readSCSITarget into a new
linux/sysfs package (with readSCSITarget exported as ReadSCSITarget) so
driver packages and the top-level linux package can share one
implementation. The subpackage location avoids an import cycle since
linux already depends on megaraid/mpi3mr/smartpqi.

Files:
- linux/sysfs.go -> linux/sysfs/sysfs.go (+package doc)
- linux/sysfs_test.go -> linux/sysfs/sysfs_test.go
- megaraid/scsi.go -> linux/sysfs/scsi.go (renamed/exported, generic
  doc covering both megaraid DID and mpi3mr PID)
- megaraid/scsi_test.go -> linux/sysfs/scsi_test.go

Callers:
- linux/system.go: default isSysPathRAID is now sysfs.IsSysPathRAID.
- megaraid/storcli.go, mpi3mr/storcli2.go: default scsiTargetFn is now
  sysfs.ReadSCSITarget. Injection type is unchanged so existing stub
  tests keep compiling.
- mpi3mr/storcli2_test.go: the one test that referenced the production
  helper directly now uses sysfs.ReadSCSITarget.
- linux/util.go: add thin IsSysPathRAID and GetSysPaths wrappers that
  forward to the sysfs package, marked // Deprecated: so existing
  external callers of the top-level linux package keep working.
