+++ linux/sysfs/sysfs_test.go
@@ -1,4 +1,4 @@
-package linux
+package sysfs
 
 import (
 	"os"
