+++ linux/system.go
@@ -12,6 +12,7 @@ import (
 
 	"golang.org/x/sys/unix"
 	"machinerun.io/disko"
+	"machinerun.io/disko/linux/sysfs"
 	"machinerun.io/disko/megaraid"
 	"machinerun.io/disko/mpi3mr"
 	"machinerun.io/disko/smartpqi"
@@ -33,7 +34,7 @@ func System() disko.System {
 			smartpqi.ArcConf(),
 			mpi3mr.StorCli2(),
 		},
-		isSysPathRAID: IsSysPathRAID,
+		isSysPathRAID: sysfs.IsSysPathRAID,
 	}
 }
